fixed some eslint errors
