+++ src/test/functionalTests/features/response/pages/defendantLipResponse/defendantDQ/welshLanguage.js
@@ -35,7 +35,7 @@ const buttons = {
 class WelshLanguage {
 
   async selectLanguageOption() {
-    const language = sharedData.language; 
+    const language = sharedData.language; 
     await I.waitForText(content.heading[language], config.WaitForText);
     await I.see(content.descriptionText[language]);
     await I.see(content.speakLanguageQuestion[language]);
