+++ src/test/functionalTests/tests/wa/welshResponse_caseworker_request_translation_tests.js
@@ -4,7 +4,6 @@ const LoginSteps = require('../../features/home/steps/login');
 const DashboardSteps = require('../../features/dashboard/steps/dashboard');
 const {unAssignAllUsers} = require('../../specClaimHelpers/api/caseRoleAssignmentHelper');
 const {createAccount, deleteAccount} = require('../../specClaimHelpers/api/idamHelper');
-const { language } = require('../../sharedData');
 
 const dontWantMoreTime = 'dontWantMoreTime';
 const bySetDate = 'bySetDate';
